+++ Classes/Domain/Model/Category.php
@@ -5,14 +5,12 @@ declare(strict_types=1);
 namespace MarekSkopal\MsReference\Domain\Model;
 
 use Stringable;
-use TYPO3\CMS\Extbase\Annotation\ORM\Lazy;
 use TYPO3\CMS\Extbase\Domain\Model\FileReference;
 use TYPO3\CMS\Extbase\DomainObject\AbstractEntity;
 use TYPO3\CMS\Extbase\Persistence\ObjectStorage;
 
 class Category extends AbstractEntity implements Stringable
 {
-    #[Lazy()]
     protected ?Category $parent = null;
 
     protected string $title = '';
@@ -30,7 +28,6 @@ class Category extends AbstractEntity implements Stringable
     protected string $text = '';
 
     /** @var ObjectStorage<FileReference> */
-    #[Lazy()]
     protected ObjectStorage $images;
 
     protected int $sorting = 0;